DESCW-901 Classes with Loader sysytem implemented
