+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.1.1: February 12, 2023
+- Restructured `src` and `inc/core` directories – to put more focus on the src directory. Moved away from functional structure to a class/loader system. Added an autoloader to the Setup and reworked the hooks system to utilise this approach through `src/Actions` and `src/Filters` class directories. Admin pages for options, notifications banner and documentation pages moved into `src/Templates`. ([DESCW-911]https://apps.itsm.gov.bc.ca/jira/browse/DESCW-901))
+
 ## 1.1.0: February 8, 2023
 - Adds a custom post type for "Custom Patterns" and register any categorised Custom Pattern post type as patterns within the pattern picker. ([DESCW-911]https://apps.itsm.gov.bc.ca/jira/browse/DESCW-911))
 - Adds the ability to include aliases or keywords that help users discover a pattern while using the pattern search. Includes admin UI updates + twitter block inclusion. ([DESCW-911]https://apps.itsm.gov.bc.ca/jira/browse/DESCW-911))
