+++ checklist.md
@@ -1,4 +1,4 @@
-Created at 2023-02-08 10:51 am
+Created at 2023-02-12 3:00 pm
 
 * [yes] Updated version in composer.json
 * [yes] Updated version in style.css or plugin file
