+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "bcgov-theme/bcgov-wordpress-block-theme",
-    "version": "1.1.0",
+    "version": "1.1.1",
     "type": "wordpress-theme",
     "repositories": [
         {
