+++ inc/core/theme-blocks-allow-list.js
@@ -1,15 +1,16 @@
 /**
-* ALLOWLIST: for Blocks in Admin
-* Manage Block Types By Post Type
-* This unfortunately is not possible in php.
-*/
-wp.domReady(function () {
+ * ALLOWLIST: for Blocks in Admin
+ * Manage Block Types By Post Type
+ * This unfortunately is not possible in php.
+ * 
+ * Note: this feature seems to require the inc/core directory.
+ */
+wp.domReady(function() {
 	/**
-	* Define the blocks we want on pages
-	*/
-	if (postData.postType == 'page' || postData.postType === 'post') {
-		wp.blocks.getBlockTypes().forEach(function (block) {
-
+	 * Define the blocks we want on pages
+	 */
+	if (postData.postType === 'page' || postData.postType === 'post') {
+		wp.blocks.getBlockTypes().forEach(function(block) {
 			// Allow by Category.
 			if (block.category === 'bcgov') return;
 
@@ -23,10 +24,17 @@ wp.domReady(function () {
 
 			// Remove Variations of core/embed.
 			if (block.name === 'core/embed') {
-				block.variations.forEach(variant => {
+				block.variations.forEach((variant) => {
 					// Allow Youtube.
-					if ('youtube' !== variant.name && 'vimeo' !== variant.name  && 'twitter' !== variant.name) {
-						wp.blocks.unregisterBlockVariation(block.name, variant.name);
+					if (
+						'youtube' !== variant.name &&
+						'vimeo' !== variant.name &&
+						'twitter' !== variant.name
+					) {
+						wp.blocks.unregisterBlockVariation(
+							block.name,
+							variant.name
+						);
 					}
 				});
 				return;
@@ -37,16 +45,12 @@ wp.domReady(function () {
 
 			// Unregister all other blocks.
 			wp.blocks.unregisterBlockType(block.name);
-		})
-	}
-
-	/**
-	* Safe defaults for custom post types.
-	*/
-	else if (postData.postType == 'cleanbc-actions') {
-
-		wp.blocks.getBlockTypes().forEach(function (block) {
-
+		});
+	} else if (postData.postType === 'cleanbc-actions') {
+		/**
+		 * Safe defaults for custom post types.
+		 */
+		wp.blocks.getBlockTypes().forEach(function(block) {
 			if (block.category === 'text') return;
 
 			if (block.category === 'media') return;
@@ -56,15 +60,22 @@ wp.domReady(function () {
 			if (block.category === 'theme') return;
 
 			if (block.name === 'core/embed') {
-				block.variations.forEach(variant => {
-					if ('youtube' !== variant.name && 'vimeo' !== variant.name && 'twitter' !== variant.name) {
-						wp.blocks.unregisterBlockVariation(block.name, variant.name);
+				block.variations.forEach((variant) => {
+					if (
+						'youtube' !== variant.name &&
+						'vimeo' !== variant.name &&
+						'twitter' !== variant.name
+					) {
+						wp.blocks.unregisterBlockVariation(
+							block.name,
+							variant.name
+						);
 					}
 				});
 				return;
 			}
 
 			wp.blocks.unregisterBlockType(block.name);
-		})
+		});
 	}
 });
