+++ src/Actions/BlockThemeSupports.php
@@ -1,0 +1,84 @@
+<?php
+
+namespace Bcgov\Theme\Block\Actions;
+
+/**
+ * Sets up theme supports.
+ *
+ * @since 1.1.1
+ *
+ * @package Bcgov/Theme/Block
+ */
+class BlockThemeSupports {
+
+	/**
+	 * Constructs theme defaults and settings.
+	 */
+	public function __construct() {
+		$this->bcgov_block_theme();
+	}
+
+	/**
+	 * Sets up theme defaults and registers support for various WordPress features.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function bcgov_block_theme(): void {
+
+		load_theme_textdomain( 'bcgov-block-theme', get_template_directory() . '/languages' );
+
+		// Enable support for Post Thumbnails on posts and pages.
+		add_theme_support( 'post-thumbnails' );
+
+		// Add support for Block Styles.
+		add_theme_support( 'wp-block-styles' );
+
+		// Add support for editor styles.
+		add_theme_support( 'editor-styles' );
+
+		// Add alignment option to toolbar options.
+		add_theme_support( 'align-wide' );
+
+		// Add support for custom logos.
+		add_theme_support( 'custom-logo' );
+
+		// Enqueue editor styles.
+		add_editor_style(
+			[
+				'./assets/css/blocks.css',
+			]
+		);
+
+		// Add support for responsive embedded content.
+		add_theme_support( 'responsive-embeds' );
+
+	}
+
+
+	/**
+	 * Filter the results of get_custom_logo() if no custom logo has been chosen by injecting a default image.
+	 * <-- wp:site-logo --> uses this result internally, and so can be easily used with or without choosing a custom logo.
+	 *
+	 * @param string $html Markup from original get_custom_logo call.
+	 * @since 1.0.0
+	 *
+	 * @return string
+	 */
+	public function bcgov_block_theme_custom_logo( $html ) {
+		// If the site owner has selected a site logo using the theme customizer, use it.
+		if ( has_custom_logo() ) {
+			return $html;
+		}
+		// Otherwise, inject a default logo markup for use in wp-site-logo blocks.
+		$image = '<img class="bcgov-logo" src="' . get_stylesheet_directory_uri() . '/assets/images/bc_gov_logo_transparent.svg">';
+		$html  = sprintf(
+			'<a href="%1$s" class="custom-logo-link" rel="home"%2$s>%3$s</a>',
+			esc_url( home_url( '/' ) ),
+			is_front_page() && ! is_paged() ? ' aria-current="page"' : '',
+			$image
+		);
+		return $html;
+	}
+}
