+++ style.css
@@ -7,7 +7,7 @@ Description: WordPress Block theme is a theme that leverages the full-site editi
 Requires at least: 5.9
 Tested up to: 6.0.1
 Requires PHP: 7.4
-Version: 1.1.0
+Version: 1.1.1
 License: Apache License Version 2.0
 License URI: LICENSE
 Text Domain: bcgov-blocks-theme
